+++ server/src/models/mod.rs
@@ -5,8 +5,8 @@ pub mod game;
 
 #[derive(Debug, Clone, Deserialize)]
 pub enum Action {
-    Attack(i32),
-    Defend(i32),
-    Heal(i32),
+    Attack(i64),
+    Defend(i64),
+    Heal(i64),
     None,
 }
